+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - `o` key binding opens the channel page in browser if the channels is the selected block.
 
+### Fixed
+- Reset the terminal properly when the app panics or an error is propagated
+
 ## [0.1.1] - 2022-05-13
 ### Fixed
 - Enable foreign key constraints in case they are disabled by default. [`029dc0c`](https://github.com/sarowish/ytsub/commit/029dc0c)